rename build.py file to fix pipeline
